% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\docType{class}
\name{trie}
\alias{trie}
\title{Conceptual class representation of a Trie.}
\description{
Package users should NEVER initialize or access attributes of this class
directly. Instead, use \code{trie_create()} for initialization and the rest
of the public functions for modification.
}
\section{Slots}{

\describe{
\item{\code{root}}{A trie node that is the root of the trie.}
}}

