% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\name{trie_add}
\alias{trie_add}
\title{Adds a single word to the trie.}
\usage{
trie_add(trie, word_to_add)
}
\arguments{
\item{trie}{A \code{trie}.}

\item{word_to_add}{The word to be added in the trie.}
}
\value{
a logical indicating that if the addition is successful or not.
Returns TRUE if word is added successfully
Returns FALSE if the word  already exists in the trie.
}
\description{
Adds a single word to the trie.
}
\examples{
trie <- trie_create()
trie_add(trie, "test")
}
