% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\name{trie_contain}
\alias{trie_contain}
\title{Search if a word is present in the trie.}
\usage{
trie_contain(trie, word)
}
\arguments{
\item{trie}{A \code{trie}.}

\item{word}{The word to be searched in the trie.}
}
\value{
a logical indicating that if the word is present or not.
Returns TRUE if the word is in the trie.
Returns FALSE if the word is not in the trie.
}
\description{
Search if a word is present in the trie.
}
\examples{
trie <- trie_create()
trie_contain(trie, "test")
}
