% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\name{trie_delete}
\alias{trie_delete}
\title{Deletes a single word from the trie.}
\usage{
trie_delete(trie, word_to_delete)
}
\arguments{
\item{trie}{A \code{trie}.}

\item{word_to_delete}{The word to be deleted from the trie.}
}
\value{
a logical indicating that if the deletion is successful or not.
Returns FALSE if the word to delete is not in the trie.
}
\description{
Deletes a single word from the trie.
}
\examples{
trie <- trie_create()
trie_delete(trie, "test")
}
