% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\name{trie_find_prefix}
\alias{trie_find_prefix}
\title{Finds all words that match the prefix in the trie.}
\usage{
trie_find_prefix(trie, prefix)
}
\arguments{
\item{trie}{A \code{trie}.}

\item{prefix}{The prefix of the words to search for.}
}
\value{
List of words matching the prefix
}
\description{
Finds all words that match the prefix in the trie.
}
\examples{
trie <- trie_create()
trie_find_prefix(trie, "be")
}
