% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasytrie.R
\docType{class}
\name{trie}
\alias{trie}
\title{Conceptual class representation of a Trie.}
\description{
Conceptual class representation of a Trie.
}
\section{Slots}{

\describe{
\item{\code{root}}{A trie node that is the root of the trie.}
}}

